Implementation using egui/eframe as windowing engine
There doesn't really seem to be a good way to do this that doesn't
integrate deeply with egui, without it being really complex. This turned
out to be not that complex. Render to an intermediate texture then
render that onto the egui interface, which is fine and could even allow
us to do post-processing in the future.